requirements.txt
opencv-python==4.13.0.92
requests==2.32.5
anthropic==0.83.0
elevenlabs==2.36.1
numpy
sounddevice
soundfile
python-dotenv
pygame-ce
torch
